Crypto.Nettle.Hash.Types

Documentation

class HashAlgorithm a where

HashAlgorithm is a class that hash algorithms will implement. generating a digest is a 3 step procedure:

hashInit to create a new context

hashUpdate to hash data

hashFinalize to extract the final digest

The final digest has hashDigestSize bytes, and the algorithm uses hashBlockSize as internal block size.

Methods

hashBlockSize :: Tagged a Int

Block size in bytes the hash algorithm operates on

hashDigestSize :: Tagged a Int

Digest size in bytes the hash algorithm returns

hashName :: Tagged a String

Name of the hash algorithm

hashInit :: a

Initialize a new context for this hash algorithm

hashUpdate :: a -> B.ByteString -> a

Update the context with bytestring, and return a new context with the updates.

hashFinalize :: a -> B.ByteString

Finalize a context and return a digest.

hashHMAC :: B.ByteString -> Tagged a KeyedHash

Use HashAlgorithm for HMAC; can use a optimized variant or the default one

hash :: HashAlgorithm a => B.ByteString -> Tagged a B.ByteString

Helper to hash a single (strict) B.ByteString in one step.

Example:

 untag (hash (fromString "abc") :: Tagged SHA256 B.ByteString)

hash' :: HashAlgorithm a => a -> B.ByteString -> B.ByteString

Untagged variant of hash; takes a (possible undefined) typed HashAlgorithm context as parameter.

Example:

 hash' (undefined :: SHA256) $ fromString "abc"

hashLazy :: HashAlgorithm a => L.ByteString -> Tagged a L.ByteString

Helper to hash a single (lazy) L.ByteString in one step.

Example:

 untag (hashLazy (fromString "abc") :: Tagged SHA256 L.ByteString)

hashLazy' :: HashAlgorithm a => a -> L.ByteString -> L.ByteString

Untagged variant of hashLazy; takes a (possible undefined) typed HashAlgorithm context as parameter.

Example:

 hashLazy' (undefined :: SHA256) $ fromString "abc"

class KeyedHashAlgorithm k s | k -> s where

KeyedHashAlgorithm is a class for keyed hash algorithms that take a key and a message to produce a digest. The most popular example is HMAC.

A KeyedHashAlgorithm will operate on a fixed key k and a state s; s will be updated as messages are added.

On start an implementation will generate the fixed key k and an initial state s from a B.ByteString key.

The state should usually be kept private.

Methods

implKeyedHashDigestSize :: Tagged k Int

Digest size in bytes the keyed hash algorithm returns

implKeyedHashInit :: B.ByteString -> (k, s)

Initialize state from a key

implKeyedHashUpdate :: k -> s -> B.ByteString -> s

Add more message data to the state

implKeyedHashFinalize :: k -> s -> B.ByteString

Produce final digest

data KeyedHash

KeyedHash hides the KeyedHashAlgorithm implementation; it contains the fixed key and the current state.

keyedHashDigestSize :: KeyedHashAlgorithm k s => Tagged k Int

Alias for implKeyedHashDigestSize

keyedHashDigestSize' :: KeyedHash -> Int

Get implKeyedHashDigestSize from a KeyedHash

keyedHashInit :: KeyedHashAlgorithm k s => B.ByteString -> Tagged k KeyedHash

keyedHashInit' :: KeyedHashAlgorithm k s => k -> B.ByteString -> KeyedHash

keyedHashInitPrivate :: KeyedHashAlgorithm k s => k -> s -> KeyedHash

keyedHashUpdate :: KeyedHash -> B.ByteString -> KeyedHash

keyedHashFinalize :: KeyedHash -> B.ByteString

keyedHash :: KeyedHashAlgorithm k s => B.ByteString -> B.ByteString -> Tagged k B.ByteString

keyedHash' :: KeyedHashAlgorithm k s => k -> B.ByteString -> B.ByteString -> B.ByteString

module Data.Tagged

data HMAC a

data HMACState a

hmacInit

Arguments

:: HashAlgorithm a

=> B.ByteString
key argument

-> Tagged a KeyedHash

hmacInit is the default implementation for hashHMAC and initializes a KeyedHash to calculate the HMAC for a message with the given key.

Example:

 let c = untag (hmacInit (fromString "secretkey") :: Tagged SHA256 KeyedHash) in keyedHashFinalize $ keyedHashUpdate c (fromString "secret message")

hmacInit' :: HashAlgorithm a => a -> B.ByteString -> KeyedHash

Untagged variant of hmacInit; takes a (possible undefined) typed HashAlgorithm context as parameter.

Example:

 keyedHashFinalize $ flip keyedHashUpdate (fromString "secret message") $ hmacInit' (undefined :: SHA256) (fromString "secretkey")

hmac

Arguments

:: HashAlgorithm a

=> B.ByteString
key argument

-> B.ByteString
message argument

-> Tagged a B.ByteString

calculate HMAC with a HashAlgorithm for a key and message

Example:

 untag (hmac (fromString "secretkey") (fromString "secret message") :: Tagged SHA256 B.ByteString)

hmac' :: HashAlgorithm a => a -> B.ByteString -> B.ByteString -> B.ByteString

Untagged variant of hmac; takes a (possible undefined) typed HashAlgorithm context as parameter.

Example:

 hmac' (undefined :: SHA256) (fromString "secretkey") (fromString "secret message")